Semana3_Backend/requirements.txt
# ============================================
# ESPE MedSafe - Backend Dependencies
# Python 3.11+
# ============================================

# Framework Web
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0

# Base de Datos
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Criptografía
cryptography==41.0.5
bcrypt==4.1.2

# Utilidades
python-dotenv==1.0.0
validators==0.22.0
python-dateutil==2.8.2

# Testing (Opcional)
pytest==7.4.4
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development (Opcional)
autopep8==2.0.4
